+++ tests/assets/esbuild/other/to_lowercase.ts
@@ -1,0 +1,3 @@
+export function toLowercase(text: string) {
+  return text.toLowerCase();
+}